requirements.txt
requests==2.26.0
python-dotenv==0.19.2
pandas==1.3.3
sqlalchemy==1.4.26